+++ .travis.yml
@@ -5,3 +5,12 @@ before_install:
   - "export DISPLAY=:99.0"
   - "sh -e /etc/init.d/xvfb start"
 script: ./tool/travis.sh
+
+# Saucelabs setup.
+env:
+  global:
+    - secure: Pk+/xlX8Oevs+WLou/1WK+22YX40x8vgfp3udqjxIehj6OS7H4pu6D0dFEd6tuPHgiQppVbB4YM3q7EFYSeJeTlZTNheF/WqZb0zkQkZMiJJM3+31qNv3HyqPzx5ornmMhskafaKjQR0euq8EeqXjwXp7beIslvGGWrygnNVn4U=
+    - secure: b04kBcfp8nNty4Uzu54GGtPsh2xSMqXYGF59i+aWJaxQNm4Hm12Z2II0DApA67MMz66Q+F6DA1oc98X6bULkY+zWv0F6DncLettmDC48eATL/ozx4azhjTzWQWDt8FwyUALSZ21iUdQYrqQJNcYN9rmf84ZuzilRLxzfVY/Kquk=
+
+addons:
+  sauce_connect: true
